+++ test/transform/resource/after-ecj/SetterDeprecated.java
@@ -1,0 +1,14 @@
+import lombok.Setter;
+class SetterDeprecated {
+  @Deprecated @Setter int annotation;
+  @Setter int javadoc;
+  public @java.lang.Deprecated @java.lang.SuppressWarnings("all") void setAnnotation(final int annotation) {
+    this.annotation = annotation;
+  }
+  public @java.lang.Deprecated @java.lang.SuppressWarnings("all") void setJavadoc(final int javadoc) {
+    this.javadoc = javadoc;
+  }
+  SetterDeprecated() {
+    super();
+  }
+}